requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
pyOpenSSL==24.0.0
psycopg2-binary==2.9.9